requirements.txt
aiogram
aiohttp
gspread
google-auth
python-dotenv==1.0.0 